requirements.txt
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Flask==2.0.1
Flask-JWT==0.3.2
Flask-JWT-Extended==4.2.1
google-api-core==1.28.0
google-api-python-client==2.6.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
peewee==3.14.4
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
schedule==1.1.0
six==1.16.0
slack-sdk==3.5.1
uritemplate==3.0.1
urllib3==1.26.5
Werkzeug==2.0.1
